Input to ft_address was earlier casted to int. Now casts to float as it should.

+++ output_hex_octal.c
@@ -12,7 +12,7 @@
 
 #include "printf.h"
 
-int		ft_address(int addr, t_nums info)
+int		ft_address(long addr, t_nums info)
 {
 	if (info.width_pos == 1)
 		ft_putnchars(' ', info.width -
